requirements.txt
# Changes must also be made in the following files:
# ./environment.yml
# ./meta.yaml
# ./pyproject.toml
# ./setup.py

cheroot
colorama
Cython
flask
flask_restful
lucit-licensing-python>=1.8.1
psutil
PySocks
requests
ujson
unicorn-fy
unicorn-binance-rest-api>=2.1.2
websocket-client
websockets==10.4
